pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "docling>=2.61.2",
    "feedparser>=6.0.12",
    "google-genai>=1.0.0",
    "hypothesis>=6.0.0",
    "markdown>=3.7.0",
    "markdownify>=0.11.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "youtube-transcript-api>=1.2.3",
]

[tool.setuptools]
packages = ["app"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
